LayoutTests/imported/w3c/web-platform-tests/preload/prefetch-accept-expected.txt


FAIL Document prefetch should send the exact Accept header as navigation assert_equals: expected "text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8" but got "*/*"
